2.1.1 Update
• Update to Minecraft 1.12.2
• The mod requires Tom’s Lib
• New circuit crafting and chipset crafting recipes
• WIP OpenComputers compatibility (Only Basic functions, no LuaObjects)
• Moved the Mapwriter part of the mod to a separate mod.
• Moved parts of the mod to my lib mod.
• JSON machine recipes for all machines
• Added JEI handler for UV Lightbox, Laser Engraver, In World Crafting
and Research Tree
• Added Open Crate
• Few texture and model changes.
• Added Crushed Ore Drops to Hammer Recipes
• Few other bug fixes

+++ src/main/java/com/tom/defense/tileentity/TileEntityForceConverter.java
@@ -1,7 +1,6 @@
 package com.tom.defense.tileentity;
 
-import static com.tom.api.energy.EnergyType.FORCE;
-import static com.tom.api.energy.EnergyType.HV;
+import static com.tom.api.energy.EnergyType.*;
 
 import java.util.List;
 
@@ -30,10 +29,10 @@ import com.tom.api.tileentity.IForcePowerStation;
 import com.tom.api.tileentity.ISecuredTileEntity;
 import com.tom.api.tileentity.ISecurityStation;
 import com.tom.api.tileentity.TileEntityTomsMod;
-import com.tom.apis.TomsModUtils;
 import com.tom.config.ConfigurationForceConverterControl;
 import com.tom.defense.ForceDeviceControlType;
 import com.tom.defense.block.ForceConverter;
+import com.tom.util.TomsModUtils;
 
 public class TileEntityForceConverter extends TileEntityTomsMod implements IEnergyHandler, IConfigurable, ISecuredTileEntity {
 	public EnergyStorage energy = new EnergyStorage(10000);
@@ -75,7 +74,7 @@ public class TileEntityForceConverter extends TileEntityTomsMod implements IEner
 	}
 
 	@Override
-	public int getMaxEnergyStored(EnumFacing from, EnergyType type) {
+	public long getMaxEnergyStored(EnumFacing from, EnergyType type) {
 		return type == FORCE ? energy.getMaxEnergyStored() : 0;
 	}
