+++ src/main/java/com/tom/factory/tileentity/TileEntityAlloySmelter.java
@@ -62,7 +62,7 @@ public class TileEntityAlloySmelter extends TileEntityMachineBase {
 		return MathHelper.ceil(energy.getEnergyStored());
 	}
 
-	public int getMaxEnergyStored() {
+	public long getMaxEnergyStored() {
 		return energy.getMaxEnergyStored();
 	}
