+++ src/main/java/com/tom/factory/tileentity/TileEntityCrusher.java
@@ -59,7 +59,7 @@ public class TileEntityCrusher extends TileEntityMachineBase {
 		return MathHelper.ceil(energy.getEnergyStored());
 	}
 
-	public int getMaxEnergyStored() {
+	public long getMaxEnergyStored() {
 		return energy.getMaxEnergyStored();
 	}
