+++ src/main/java/com/tom/factory/tileentity/TileEntityElectricFurnace.java
@@ -66,7 +66,7 @@ public class TileEntityElectricFurnace extends TileEntityMachineBase {
 		return MathHelper.ceil(energy.getEnergyStored());
 	}
 
-	public int getMaxEnergyStored() {
+	public long getMaxEnergyStored() {
 		return energy.getMaxEnergyStored();
 	}
