+++ src/main/java/com/tom/factory/tileentity/TileEntityPlateBlendingMachine.java
@@ -47,14 +47,12 @@ public class TileEntityPlateBlendingMachine extends TileEntityMachineBase {
 	public void readFromNBT(NBTTagCompound compound) {
 		super.readFromNBT(compound);
 		this.progress = compound.getInteger("progress");
-		// this.maxProgress = compound.getInteger("maxProgress");
 	}
 
 	@Override
 	public NBTTagCompound writeToNBT(NBTTagCompound compound) {
 		super.writeToNBT(compound);
 		compound.setInteger("progress", progress);
-		// compound.setInteger("mayProgress", maxProgress);
 		return compound;
 	}
 
@@ -62,7 +60,7 @@ public class TileEntityPlateBlendingMachine extends TileEntityMachineBase {
 		return MathHelper.ceil(energy.getEnergyStored());
 	}
 
-	public int getMaxEnergyStored() {
+	public long getMaxEnergyStored() {
 		return energy.getMaxEnergyStored();
 	}
